+++ spec/app/models/wiki_page.rb
@@ -11,6 +11,6 @@ class WikiPage
   attr_protected :author
 
   has_many :comments, :dependent => :destroy, :validate => false
-  has_many :child_pages, :class_name => "WikiPage", :dependent => :delete
-  belongs_to :parent_pages, :class_name => "WikiPage"
+  has_many :child_pages, :class_name => "WikiPage", :dependent => :delete, :inverse_of => :parent_pages
+  belongs_to :parent_pages, :class_name => "WikiPage", :inverse_of => :child_pages
 end